Wrap rendering with delegate
