+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/track_selection/details/fragment/TrackSelectionDetailsFragment.kt
@@ -2,33 +2,23 @@ package org.hyperskill.app.android.track_selection.details.fragment
 
 import android.os.Bundle
 import android.view.View
-import android.view.ViewGroup.MarginLayoutParams
-import android.widget.TextView
-import androidx.core.view.doOnNextLayout
-import androidx.core.view.isVisible
-import androidx.core.view.marginBottom
-import androidx.core.view.updateLayoutParams
-import androidx.core.view.updateMargins
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import by.kirich1409.viewbindingdelegate.viewBinding
 import org.hyperskill.app.android.HyperskillApp
 import org.hyperskill.app.android.R
 import org.hyperskill.app.android.core.extensions.argument
-import org.hyperskill.app.android.core.view.ui.dialog.LoadingProgressDialogFragment
-import org.hyperskill.app.android.core.view.ui.dialog.dismissDialogFragmentIfExists
 import org.hyperskill.app.android.core.view.ui.navigation.requireRouter
 import org.hyperskill.app.android.databinding.FragmentTrackSelectionDetailsBinding
 import org.hyperskill.app.android.main.view.ui.navigation.MainScreenRouter
 import org.hyperskill.app.android.study_plan.screen.StudyPlanScreen
+import org.hyperskill.app.android.track_selection.details.delegate.TrackSelectionDetailsDelegate
 import org.hyperskill.app.track_selection.details.injection.TrackSelectionDetailsParams
 import org.hyperskill.app.track_selection.details.presentation.TrackSelectionDetailsFeature.Action.ViewAction
 import org.hyperskill.app.track_selection.details.presentation.TrackSelectionDetailsFeature.Message
 import org.hyperskill.app.track_selection.details.presentation.TrackSelectionDetailsFeature.ViewState
 import org.hyperskill.app.track_selection.details.presentation.TrackSelectionDetailsViewModel
 import ru.nobird.android.view.base.ui.delegate.ViewStateDelegate
-import ru.nobird.android.view.base.ui.extension.setTextIfChanged
-import ru.nobird.android.view.base.ui.extension.showIfNotExists
 import ru.nobird.android.view.base.ui.extension.snackbar
 import ru.nobird.android.view.redux.ui.extension.reduxViewModel
 import ru.nobird.app.presentation.redux.container.ReduxView
@@ -122,68 +112,12 @@ class TrackSelectionDetailsFragment :
     override fun render(state: ViewState) {
         viewStateDelegate?.switchState(state)
         if (state is ViewState.Content) {
-            updateContentBottomMargin()
-            renderSelectionLoading(state.isTrackSelectionLoadingShowed)
-            viewBinding.projectSelectionDetailsToolbar.title = state.title
-            with(viewBinding.trackSelectionDetailsDescription) {
-                setNullableText(trackSelectionDetailsDescription, state.description)
-                trackSelectionDetailsDescription.updateLayoutParams<MarginLayoutParams> {
-                    updateMargins(
-                        top = if (state.areTagsVisible) {
-                            resources.getDimensionPixelOffset(R.dimen.track_selection_details_tags_bottom_margin)
-                        } else {
-                            0
-                        }
-                    )
-                }
-                trackSelectionDetailsSelectedBadge.root.isVisible = state.isSelected
-                trackSelectionDetailsBetaBadge.root.isVisible = state.isBeta
-                trackSelectionDetailsCompletedBadge.root.isVisible = state.isCompleted
-            }
-            with(viewBinding.trackSelectionDetailsTrackOverview) {
-                trackOverviewRatingTextView.setTextIfChanged(state.formattedRating)
-                setNullableText(trackOverviewHoursToCompleteTextView, state.formattedTimeToComplete)
-                trackOverviewTopicsCountTextView.setTextIfChanged(state.formattedTopicsCount)
-                setNullableText(trackOverviewProjectsCountTextView, state.formattedProjectsCount)
-                trackOverviewCertificateAvailabilityTextView.isVisible = state.isCertificateAvailable
-            }
-            with(viewBinding.trackSelectionDetailsProviders) {
-                setNullableText(trackSelectionDetailsMainProviderTitle, state.mainProvider?.title)
-                setNullableText(trackSelectionDetailsMainProviderDescription, state.mainProvider?.description)
-                trackSelectionDetailsMainProviderDivider.isVisible = state.formattedOtherProviders != null
-                trackSelectionDetailsOtherProvidersTitle.isVisible = state.formattedOtherProviders != null
-                setNullableText(trackSelectionDetailsOtherProvidersDescription, state.formattedOtherProviders)
-            }
-        }
-    }
-
-    private fun renderSelectionLoading(isLoadingShowed: Boolean) {
-        if (isLoadingShowed) {
-            LoadingProgressDialogFragment.newInstance()
-                .showIfNotExists(childFragmentManager, LoadingProgressDialogFragment.TAG)
-        } else {
-            childFragmentManager.dismissDialogFragmentIfExists(LoadingProgressDialogFragment.TAG)
-        }
-    }
-
-    private fun setNullableText(textView: TextView, text: String?) {
-        textView.isVisible = text != null
-        if (text != null) {
-            textView.setTextIfChanged(text)
-        }
-    }
-
-    /**
-     * Add bottom margin to content container to prevent overlapping with select button
-     */
-    private fun updateContentBottomMargin() {
-        viewBinding.trackSelectionDetailsSelectButton.doOnNextLayout {
-            viewBinding.trackSelectionDetailsContentContainer.updateLayoutParams<MarginLayoutParams> {
-                updateMargins(
-                    bottom = it.height + it.marginBottom +
-                        resources.getDimensionPixelOffset(R.dimen.track_selection_details_content_bottom_margin)
-                )
-            }
+            TrackSelectionDetailsDelegate.render(
+                context = requireContext(),
+                viewBinding = viewBinding,
+                fragmentManager = childFragmentManager,
+                state = state
+            )
         }
     }
 }
